requirements.txt
torch
torchvision
matplotlib
scikit-learn
pillow
seaborn
